+++ internal/dnsforward/msg.go
@@ -1,7 +1,7 @@
 package dnsforward
 
 import (
-	"net"
+	"net/netip"
 	"time"
 
 	"github.com/AdguardTeam/AdGuardHome/internal/filtering"
@@ -9,6 +9,7 @@ import (
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/AdguardTeam/urlfilter/rules"
 	"github.com/miekg/dns"
+	"golang.org/x/exp/slices"
 )
 
 // makeResponse creates a DNS response by req and sets necessary flags.  It also
@@ -26,24 +27,13 @@ func (s *Server) makeResponse(req *dns.Msg) (resp *dns.Msg) {
 	return resp
 }
 
-// containsIP returns true if the IP is already in the list.
-func containsIP(ips []net.IP, ip net.IP) bool {
-	for _, a := range ips {
-		if a.Equal(ip) {
-			return true
-		}
-	}
-
-	return false
-}
-
 // ipsFromRules extracts unique non-IP addresses from the filtering result
 // rules.
-func ipsFromRules(resRules []*filtering.ResultRule) (ips []net.IP) {
+func ipsFromRules(resRules []*filtering.ResultRule) (ips []netip.Addr) {
 	for _, r := range resRules {
 		// len(resRules) and len(ips) are actually small enough for O(n^2) to do
 		// not raise performance questions.
-		if ip := r.IP; ip != nil && !containsIP(ips, ip) {
+		if ip := r.IP; ip != (netip.Addr{}) && !slices.Contains(ips, ip) {
 			ips = append(ips, ip)
 		}
 	}
@@ -84,7 +74,7 @@ func (s *Server) genDNSFilterMessage(
 
 // genForBlockingMode generates a filtered response to req based on the server's
 // blocking mode.
-func (s *Server) genForBlockingMode(req *dns.Msg, ips []net.IP) (resp *dns.Msg) {
+func (s *Server) genForBlockingMode(req *dns.Msg, ips []netip.Addr) (resp *dns.Msg) {
 	qt := req.Question[0].Qtype
 	switch m := s.conf.BlockingMode; m {
 	case BlockingModeCustomIP:
@@ -126,13 +116,13 @@ func (s *Server) genServerFailure(request *dns.Msg) *dns.Msg {
 	return &resp
 }
 
-func (s *Server) genARecord(request *dns.Msg, ip net.IP) *dns.Msg {
+func (s *Server) genARecord(request *dns.Msg, ip netip.Addr) *dns.Msg {
 	resp := s.makeResponse(request)
 	resp.Answer = append(resp.Answer, s.genAnswerA(request, ip))
 	return resp
 }
 
-func (s *Server) genAAAARecord(request *dns.Msg, ip net.IP) *dns.Msg {
+func (s *Server) genAAAARecord(request *dns.Msg, ip netip.Addr) *dns.Msg {
 	resp := s.makeResponse(request)
 	resp.Answer = append(resp.Answer, s.genAnswerAAAA(request, ip))
 	return resp
@@ -147,17 +137,17 @@ func (s *Server) hdr(req *dns.Msg, rrType rules.RRType) (h dns.RR_Header) {
 	}
 }
 
-func (s *Server) genAnswerA(req *dns.Msg, ip net.IP) (ans *dns.A) {
+func (s *Server) genAnswerA(req *dns.Msg, ip netip.Addr) (ans *dns.A) {
 	return &dns.A{
 		Hdr: s.hdr(req, dns.TypeA),
-		A:   ip,
+		A:   ip.AsSlice(),
 	}
 }
 
-func (s *Server) genAnswerAAAA(req *dns.Msg, ip net.IP) (ans *dns.AAAA) {
+func (s *Server) genAnswerAAAA(req *dns.Msg, ip netip.Addr) (ans *dns.AAAA) {
 	return &dns.AAAA{
 		Hdr:  s.hdr(req, dns.TypeAAAA),
-		AAAA: ip,
+		AAAA: ip.AsSlice(),
 	}
 }
 
@@ -204,22 +194,24 @@ func (s *Server) genAnswerTXT(req *dns.Msg, strs []string) (ans *dns.TXT) {
 // addresses and an appropriate resource record type.  If any of the IPs cannot
 // be converted to the correct protocol, genResponseWithIPs returns an empty
 // response.
-func (s *Server) genResponseWithIPs(req *dns.Msg, ips []net.IP) (resp *dns.Msg) {
+func (s *Server) genResponseWithIPs(req *dns.Msg, ips []netip.Addr) (resp *dns.Msg) {
 	var ans []dns.RR
 	switch req.Question[0].Qtype {
 	case dns.TypeA:
 		for _, ip := range ips {
-			if ip4 := ip.To4(); ip4 == nil {
+			if ip.Is4() {
+				ans = append(ans, s.genAnswerA(req, ip))
+			} else {
 				ans = nil
 
 				break
 			}
-
-			ans = append(ans, s.genAnswerA(req, ip))
 		}
 	case dns.TypeAAAA:
 		for _, ip := range ips {
-			ans = append(ans, s.genAnswerAAAA(req, ip.To16()))
+			if ip.Is6() {
+				ans = append(ans, s.genAnswerAAAA(req, ip))
+			}
 		}
 	default:
 		// Go on and return an empty response.
@@ -240,9 +232,9 @@ func (s *Server) makeResponseNullIP(req *dns.Msg) (resp *dns.Msg) {
 	// converted into an empty slice instead of the zero IPv4.
 	switch req.Question[0].Qtype {
 	case dns.TypeA:
-		resp = s.genResponseWithIPs(req, []net.IP{{0, 0, 0, 0}})
+		resp = s.genResponseWithIPs(req, []netip.Addr{netip.IPv4Unspecified()})
 	case dns.TypeAAAA:
-		resp = s.genResponseWithIPs(req, []net.IP{net.IPv6zero})
+		resp = s.genResponseWithIPs(req, []netip.Addr{netip.IPv6Unspecified()})
 	default:
 		resp = s.makeResponse(req)
 	}
@@ -251,9 +243,9 @@ func (s *Server) makeResponseNullIP(req *dns.Msg) (resp *dns.Msg) {
 }
 
 func (s *Server) genBlockedHost(request *dns.Msg, newAddr string, d *proxy.DNSContext) *dns.Msg {
-	ip := net.ParseIP(newAddr)
-	if ip != nil {
-		return s.genResponseWithIPs(request, []net.IP{ip})
+	ip, err := netip.ParseAddr(newAddr)
+	if err == nil {
+		return s.genResponseWithIPs(request, []netip.Addr{ip})
 	}
 
 	// look up the hostname, TODO: cache
@@ -275,7 +267,7 @@ func (s *Server) genBlockedHost(request *dns.Msg, newAddr string, d *proxy.DNSCo
 		return s.genServerFailure(request)
 	}
 
-	err := prx.Resolve(newContext)
+	err = prx.Resolve(newContext)
 	if err != nil {
 		log.Printf("couldn't look up replacement host %q: %s", newAddr, err)